fix(mojo): update source files for Mojo nightly compatibility
- 01_advanced_traits: MutOrigin.external → MutExternalOrigin; add
  ImplicitlyDestructible to GenericArray and Container constraints
- 01_ctx_mgr: fix SIMD division by casting to Float64 before dividing by 1e6
- 01_types: comment out emberjson usage (binary incompatible with current nightly)
- 01_working_with_traits: replace pass with ... in trait default method bodies
- 01_parameterization: *_ wildcard → ... in partially-bound type parameter
- 01_pointers: rename __copyinit__ arg other → copy per new Mojo requirement
- pixi.toml: add pygame via conda-forge for 01_helloworld

+++ 01_basics/01_ctx_mgr.mojo
@@ -30,7 +30,7 @@ struct Timer(ImplicitlyCopyable, Movable):
 
   fn __exit__(mut self):
     end_time = time.perf_counter_ns()
-    elapsed_time_ms = round(((end_time - UInt(self.start_time)) / 1e6), 3)
+    elapsed_time_ms = round((Float64(end_time) - Float64(self.start_time)) / 1e6, 3)
     print("Elapsed time: ", elapsed_time_ms, " milliseconds")
 
 @fieldwise_init
@@ -46,7 +46,7 @@ struct ConditionalTimer(ImplicitlyCopyable, Movable):
 
   fn __exit__(mut self):
     end_time = time.perf_counter_ns()
-    elapsed_time_ms = round(((end_time - UInt(self.start_time)) / 1e6), 3)
+    elapsed_time_ms = round((Float64(end_time) - Float64(self.start_time)) / 1e6, 3)
     print("Elapsed time: ", elapsed_time_ms, " milliseconds")
 
   fn __exit__(mut self, e : Error) raises -> Bool:
